+++ FirebaseAdmin/FirebaseAdmin/Auth/FirebaseTokenVerifierArgs.cs
@@ -31,5 +31,20 @@ namespace FirebaseAdmin.Auth
         public IClock Clock { get; set; }
 
         public IPublicKeySource PublicKeySource { get; set; }
+
+        internal static FirebaseTokenVerifierArgs CreateIdTokenVerififierArgs(
+            string projectId, IPublicKeySource keySource, IClock clock = null)
+        {
+            return new FirebaseTokenVerifierArgs()
+            {
+                ProjectId = projectId,
+                ShortName = "ID token",
+                Operation = "VerifyIdTokenAsync()",
+                Url = "https://firebase.google.com/docs/auth/admin/verify-id-tokens",
+                Issuer = "https://securetoken.google.com/",
+                Clock = clock ?? SystemClock.Default,
+                PublicKeySource = keySource,
+            };
+        }
     }
 }
